requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Data processing
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4

# Machine learning and dimensionality reduction
umap-learn==0.5.5
scikit-learn==1.3.2

# Network analysis
networkx==3.2.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Statistical analysis
statsmodels==0.14.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# LLM Support (optional)
openai==1.3.0
anthropic==0.7.0

# Enhanced ML
scikit-learn==1.3.2
